Move solutions scanning logic into stand-alone functions organized by target page.
