+++ tsconfig.json
@@ -32,7 +32,9 @@
       "@app/solutions-scanner": ["libs/solutions-scanner/src"],
       "@app/solutions-scanner/*": ["libs/solutions-scanner/src/*"],
       "@app/storage": ["libs/storage/src"],
-      "@app/storage/*": ["libs/storage/src/*"]
+      "@app/storage/*": ["libs/storage/src/*"],
+      "@app/page-scanners": ["libs/page-scanners/src"],
+      "@app/page-scanners/*": ["libs/page-scanners/src/*"]
     }
   }
 }
